Move seed function to separate file
